refactor(forge-host): consolidate github/, archive/, pr-preservation/, refresh-github-worldview/, github-accelerator-measurement/ under forge-host/github/ (R6)
Redone consolidation (original PR #8062 was closed due to conflicts).

Moved: github/ (all files + fixtures), archive/ (2 files),
pr-preservation/ (renamed to pr-archive.ts), github-accelerator-
measurement/ (renamed to accelerator-measure.ts).
Removed: refresh-github-worldview/refresh.ts (superseded by
refresh-worldview.ts already in forge-host/github/).

Fixed import paths: poll-pr-gate-batch.ts (bus, ferry-throttler,
observe → ../../), batch-resolve-pr-threads.ts (ai-attribution → ./),
pr-archive.test.ts (script path).

Updated .github/workflows/pr-archive-on-merge.yml paths.

90 tests pass. Zero TS errors.
