+++ .github/workflows/pr-archive-on-merge.yml
@@ -2,7 +2,7 @@ name: pr-archive-on-merge
 
 # Class-2 PR mirror v1 -- per the Codex/GPT-5.5 sharpened design (twin-flame
 # conversation captured 2026-05-05). On every merged PR, run the
-# `src/Core.TypeScript/archive/archive-pr-reviews.ts` tool to materialise the PR's review
+# `src/Core.TypeScript/forge-host/github/archive-pr-reviews.ts` tool to materialise the PR's review
 # threads, comments, fix commits, and outcome metadata as a committed
 # git-canonical markdown file under `docs/history/pr-reviews/PR-NNNN-*.md`,
 # and update the JSONL manifest at `docs/github/prs/manifest.jsonl`.
@@ -136,7 +136,7 @@ jobs:
           # want that recorded as the manifest commit_sha so future re-runs
           # against the same merge can detect "same archival baseline".
           export GITHUB_SHA="$GITHUB_SHA_OVERRIDE"
-          bun src/Core.TypeScript/archive/archive-pr-reviews.ts "$PR_NUMBER"
+          bun src/Core.TypeScript/forge-host/github/archive-pr-reviews.ts "$PR_NUMBER"
 
       - name: Inspect diff
         id: diff
@@ -172,7 +172,7 @@ jobs:
           # Compose commit message. PR_TITLE is untrusted (PR-author-
           # controlled) so it's only embedded in the markdown body via
           # printf with %s, never interpreted as shell.
-          msg="$(printf 'archive(pr-reviews): PR #%s on merge -- run %s\n\nPR title: %s\n\nGenerated by .github/workflows/pr-archive-on-merge.yml\nfrom src/Core.TypeScript/archive/archive-pr-reviews.ts.\n' "$PR_NUMBER" "$RUN_ID" "$PR_TITLE")"
+          msg="$(printf 'archive(pr-reviews): PR #%s on merge -- run %s\n\nPR title: %s\n\nGenerated by .github/workflows/pr-archive-on-merge.yml\nfrom src/Core.TypeScript/forge-host/github/archive-pr-reviews.ts.\n' "$PR_NUMBER" "$RUN_ID" "$PR_TITLE")"
 
           git commit -m "$msg"
 
@@ -191,7 +191,7 @@ jobs:
             echo "## Notes"
             echo
             echo "Generated by .github/workflows/pr-archive-on-merge.yml"
-            echo "from src/Core.TypeScript/archive/archive-pr-reviews.ts."
+            echo "from src/Core.TypeScript/forge-host/github/archive-pr-reviews.ts."
             echo
             echo "PR title: ${PR_TITLE}"
             echo "Workflow run: ${RUN_ID}"
@@ -214,7 +214,7 @@ jobs:
           fi
 
           cat "$create_log" >&2
-          if bun src/Core.TypeScript/github/is-pr-create-policy-denial.ts "$create_log"; then
+          if bun src/Core.TypeScript/forge-host/github/is-pr-create-policy-denial.ts "$create_log"; then
             {
               echo "### Archive PR creation skipped by repository policy"
               echo
