+++ src/Core.TypeScript/forge-host/github/fixtures/clean-armed-auto-merge.json
